More refactoring, and one real change.
pywintypesXX.dll and pythoncomXX.dll are now copied into the target
exe directory. While I'm not sure that this is really sufficient for
COM, it's far better than before. They would not be found in the lib
directory for inproc COM servers at least.

Still when a (possibly conflicting) Python and win32all installation
is on the same machine, the wrong dlls are loaded.

See comment in the code.